front: move getSavedMacroNodes() to osrdEditoastApi
Move it alongside the rest of the paginated helpers.
